feature update: booking update
card UI and booking detail modal update
