+++ hamiguitan_trekscan_plus/android/app/src/main/AndroidManifest.xml
@@ -29,7 +29,8 @@
         android:label="Hamiguitan TrekScan+"
         android:name="${applicationName}"
         android:icon="@mipmap/launcher_icon"
-        android:requestLegacyExternalStorage="true">
+        android:requestLegacyExternalStorage="true"
+        android:enableOnBackInvokedCallback="true">
         <activity
             android:name=".MainActivity"
             android:exported="true"
